Add AppArtifactConfigCommand and unit test. Update test names to be more descriptive and break out maven versus overall tests.
