+++ liberty-starter-application/skeletons/specialFiles/pom.xml
@@ -3,8 +3,8 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>myParentGroupId</groupId>
-    <artifactId>myArtifactId</artifactId>
+    <groupId>liberty.maven</groupId>
+    <artifactId>test</artifactId>
     <version>1.0-SNAPSHOT</version>
     <packaging>war</packaging>
